Move nodes definition into a new file
